refactor(dates): generate display dates in container components - app.tsx for now - will split this up - titleBar comp is display only now - this allows the link data to be synced
